feat(pattern-generator): adds basic stitch grid generator
This commit adds a StitchGrid vue component that takes in a height
and width and generates a cross stitch grid of that size. It allows
customization of the size of each square on the pattern grid for easy
reading. It also has basic functionality to fill a pattern into the grid.